[FEAT] add movement cooldown
