+++ src/features/game/events/landExpansion/moveCollectible.test.ts
@@ -12,6 +12,7 @@ const GAME_STATE: GameState = {
 };
 
 describe("moveCollectible", () => {
+  const dateNow = Date.now();
   it("throws if player has no Bumpkin", () => {
     expect(() =>
       moveCollectible({
@@ -80,7 +81,7 @@ describe("moveCollectible", () => {
     ).toThrow(MOVE_COLLECTIBLE_ERRORS.COLLECTIBLE_NOT_PLACED);
   });
 
-  it("moves a building", () => {
+  it("moves a collectible", () => {
     const gameState = moveCollectible({
       state: {
         ...GAME_STATE,
@@ -116,6 +117,128 @@ describe("moveCollectible", () => {
     });
 
     expect(gameState.collectibles["Abandoned Bear"]).toEqual([
+      {
+        id: "123",
+        coordinates: { x: 2, y: 2 },
+        createdAt: 0,
+        readyAt: 0,
+      },
+      {
+        id: "456",
+        coordinates: { x: 4, y: 4 },
+        createdAt: 0,
+        readyAt: 0,
+      },
+      {
+        id: "789",
+        coordinates: {
+          x: 8,
+          y: 8,
+        },
+        createdAt: 0,
+        readyAt: 0,
+      },
+    ]);
+  });
+
+  it("updates readyAt when moving Basic Scarecrow", () => {
+    const gameState = moveCollectible({
+      state: {
+        ...GAME_STATE,
+        collectibles: {
+          "Basic Scarecrow": [
+            {
+              id: "123",
+              coordinates: { x: 1, y: 1 },
+              createdAt: 0,
+              readyAt: 0,
+            },
+            {
+              id: "456",
+              coordinates: { x: 4, y: 4 },
+              createdAt: 0,
+              readyAt: 0,
+            },
+            {
+              id: "789",
+              coordinates: { x: 8, y: 8 },
+              createdAt: 0,
+              readyAt: 0,
+            },
+          ],
+        },
+      },
+      action: {
+        type: "collectible.moved",
+        name: "Basic Scarecrow",
+        id: "123",
+        coordinates: { x: 2, y: 2 },
+      },
+      createdAt: dateNow,
+    });
+
+    expect(gameState.collectibles["Basic Scarecrow"]).toEqual([
+      {
+        id: "123",
+        coordinates: { x: 2, y: 2 },
+        createdAt: 0,
+        readyAt: dateNow + 10 * 60 * 1000,
+      },
+      {
+        id: "456",
+        coordinates: { x: 4, y: 4 },
+        createdAt: 0,
+        readyAt: 0,
+      },
+      {
+        id: "789",
+        coordinates: {
+          x: 8,
+          y: 8,
+        },
+        createdAt: 0,
+        readyAt: 0,
+      },
+    ]);
+  });
+
+  it("does not updates readyAt when moving Nancy", () => {
+    const gameState = moveCollectible({
+      state: {
+        ...GAME_STATE,
+        collectibles: {
+          Nancy: [
+            {
+              id: "123",
+              coordinates: { x: 1, y: 1 },
+              createdAt: 0,
+              readyAt: 0,
+            },
+            {
+              id: "456",
+              coordinates: { x: 4, y: 4 },
+              createdAt: 0,
+              readyAt: 0,
+            },
+            {
+              id: "789",
+              coordinates: { x: 8, y: 8 },
+              createdAt: 0,
+              readyAt: 0,
+            },
+          ],
+        },
+      },
+      action: {
+        type: "collectible.moved",
+        name: "Nancy",
+        id: "123",
+        coordinates: { x: 2, y: 2 },
+      },
+      createdAt: dateNow,
+    });
+
+    expect(gameState.collectibles["Nancy"]).toEqual([
       {
         id: "123",
         coordinates: { x: 2, y: 2 },
